feat: Implement ExecutionRepository and track step executions

+++ app/src/main/java/com/mrdarip/stepy/data/local/entities/ExecutionEntity.kt
@@ -33,6 +33,6 @@ data class ExecutionEntity(
     val parentExecutionId: Int?,
     val stepId: Int,
     val position: Int,
-    val start: Int,
-    val end: Int,
+    val start: Long,
+    val end: Long,
 )
